api/requirements.txt
langserve[server]
fastapi
langchain
langchain-community
uvicorn
langchain-google-genai
langchain-openai
pydantic
python-dotenv
httpx==0.27.2
beanie
pymongo
motor
pydantic[email]